--- source/horoscope/css/fortunePage.css
@@ -1,194 +1,0 @@
-/*
-* CSS file for fortunePage.html
-*/
-
-@import "pastEntryCard.css";
-
-/* New fonts for sidebar and input labels */
-@font-face {
-    font-family: 'montserrat';
-    src: url('fonts/Montserrat-Regular.ttf') format('woff');
-    font-weight: normal;
-    font-style: normal;
-}
-@font-face {
-    font-family: 'pacifico';
-    src: url('fonts/Pacifico-Regular.ttf') format('woff');
-    font-weight: normal;
-    font-style: normal;
-}
-
-
-/* ELEMENT SELECTORS */
-/* wrapper over all components of the webpage */
-body {
-    font-family: Garamond, Baskerville, Baskerville Old Face, Hoefler Text, Times New Roman, serif;
-    color: white;
-
-    /* takes up the full page */
-    margin: 0;
-    padding: 0;
-    height: 100%;
-    width: 100%;
-
-    display: flex;
-    flex-direction: row;
-    height: 100vh;
-}
-
-/* Video that plays when birthdate is submitted */
-video {
-    /* takes up the full page */
-    position: absolute;
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 100%;
-    object-fit: cover;
-    z-index: -1;
-}
-
-/* The center body, not including the sidebar */
-
-main {
-    /* Center and scale the image nicely */
-    background-position: center;
-    background-repeat: no-repeat;
-    background-size: cover;
-    background-origin: padding-box;
-
-    /* takes up the full page, but adjusts based on display size */
-    margin: 0;
-    padding: 0;
-    height: 100%;
-    width: 100%;
-    display: flex;
-    flex-direction: column;
-    position: relative;
-}
-
-/* ID SELECTORS */
-
-/* Background image */
-#vbgimg {
-    /* takes up the full page */
-    position: absolute;
-    object-fit: cover;
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 100%;
-    z-index: -10;
-}
-
-/* all the input related components. 'Birthday', 'Select a category', and 'Save' button. */
-#user-form {
-    /* positioning on the webpage based on bottom and left side */
-    padding-bottom: 20px;
-    margin-left: 25px;
-    text-align: center;
-}
-
-/* in Output */
-/* for the prompt that is displayed based on the input */
-#output {
-    /* positioning on the webpage, changes based on the webpage size but keeps itself centered */
-    position: absolute;
-    bottom: 4vw;
-    left: 50%;
-    transform: translateX(-50%);
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    font-size: 20px;
-    text-align: center;
-    width: 80%;
-}
-
-/* Name of the corresponding horoscope */
-#horoscope-title {
-    position: fixed;
-    /* Set the position property to relative */
-
-    top: 40px;
-    /* Adjust the top property to move the element vertically */
-
-    font-family: 'pacifico';
-
-    z-index: 9999; /*bring elem forward*/
-}
-
-/* Fortune prompt */
-#horoscope-fortune {
-    /* positioning in the 'output' box. text is centered at all times. */
-    display: block;
-    margin-left: auto;
-    margin-right: auto;
-    width: 80%;
-    flex: 1;
-    font-size: clamp(20px, 3vw, 30px);
-    padding: 30px;
-    padding-top: 70px;
-    border-radius: 20px;
-    background-image: linear-gradient(rgba(0, 17, 85, 0.75), rgba(0, 183, 255, 0.0));
-    position: relative;
-    top: 70px;
-}
-
-/* Save and Redo Button Styles */
-.button-container {
-    display: flex;
-    justify-content: space-around; /* Distribute buttons evenly */
-    align-items: center;
-    margin-top: 50px;
-}
-
-/* Save Button*/
-#save {
-    position: static;
-    transition: background-color 1s ease; /* change of color takes a second(ex. when you hover) */
-    background-color: #1565C0;
-    /*navy blue*/
-    color: white;
-    padding: 10px 20px;
-    border: none;
-    border-radius: 4px;
-    cursor: pointer;
-    font-size: clamp(15px, 1.5vw, 30px);
-}
-
-/* Redo Button*/
-#redo {
-    position: static;
-    transition: background-color 1s ease; /* change of color takes a second(ex. when you hover) */
-    background-color: #1565C0;
-    /*navy blue*/
-    color: white;
-    padding: 10px 20px;
-    border: none;
-    border-radius: 4px;
-    cursor: pointer;
-    font-size: clamp(15px, 1.5vw, 30px);
-}
-
-/* save and redo button when hovered by mouse */
-#save:hover,
-#redo:hover {
-    background-color: #0D47A1;
-    /*midnight blue*/
-}
-
-/* save and redo button when clicked */
-#save:active,
-#redo:active {
-    background-color: #2196F3;
-    /*light blue*/
-}
-
-/* save and redo button when 'tab' to */
-#save:focus,
-#redo:active {
-    outline: none;
-    box-shadow: 0 0 0 2px #1A237E;
-    /*dark blue*/
-}
