+++ source/horoscope/css/horoscope.css
@@ -1,10 +1,10 @@
 /*
-* CSS file for horoscope.html
+* CSS file for fortunePage.html
 */
 
 @import "pastEntryCard.css";
 
-/* New fonts for history and input labels */
+/* New fonts for sidebar and input labels */
 @font-face {
     font-family: 'montserrat';
     src: url('fonts/Montserrat-Regular.ttf') format('woff');
@@ -48,7 +48,8 @@ video {
     z-index: -1;
 }
 
-/* The center body */
+/* The center body, not including the sidebar */
+
 main {
     /* Center and scale the image nicely */
     background-position: center;
@@ -80,7 +81,6 @@ main {
     z-index: -10;
 }
 
-
 /* all the input related components. 'Birthday', 'Select a category', and 'Save' button. */
 #user-form {
     /* positioning on the webpage based on bottom and left side */
@@ -103,6 +103,7 @@ main {
     font-size: 20px;
     text-align: center;
     width: 80%;
+    margin-bottom: 2rem;
 }
 
 /* Name of the corresponding horoscope */
@@ -110,7 +111,7 @@ main {
     position: fixed;
     /* Set the position property to relative */
 
-    top: 60px;
+    top: 40px;
     /* Adjust the top property to move the element vertically */
 
     font-family: 'pacifico';
@@ -128,70 +129,24 @@ main {
     flex: 1;
     font-size: clamp(20px, 3vw, 30px);
     padding: 30px;
-    padding-top: 50px;
+    padding-top: 70px;
     border-radius: 20px;
     background-image: linear-gradient(rgba(0, 17, 85, 0.75), rgba(0, 183, 255, 0.0));
     position: relative;
     top: 70px;
 }
 
-/* 'Birthday' label */
-label[for="birthday"] {
-    font-size: 32px;
-    height: 40px;
-    line-height: 40px;
+/* Save and Redo Button Styles */
+.button-container {
+    display: flex;
+    justify-content: space-around; /* Distribute buttons evenly */
+    align-items: center;
+    margin-top: 50px;
 }
 
-/* 'Select a catergory' label */
-label[for="category"] {
-    font-size: 32px;
-    height: 40px;
-    line-height: 40px;
-}
-
-/* Birthday input */
-/* look of the birthday display on the webpage. Transparent and white to make it match the color-scheme
-    without blocking too much of the webpage */
-input[type="date"] {
-    background-color: transparent;
-    color: white;
-    border: 2px;
-    border-style: solid;
-    border-radius: 6px;
-    font-size: 24px;
-    margin-right: 35px;
-    padding: 10px 18px;
-}
-
-/* Calendar display */
-::-webkit-calendar-picker-indicator {
-    background-color: white;
-    border-radius: 6px;
-}
-
-/* 'Select a category' default */
-/* look of the category display on the webpage. Matches the visual of birthday input */
-select {
-    background-color: transparent;
-    color: white;
-    border: 2px;
-    border-style: solid;
-    border-radius: 6px;
-    font-size: clamp(20px, 3vw, 40px);
-    text-align: center;
-}
-
-/* Dropdown for select a category */
-/* Couldn't do transparent background so white background(default) with black text */
-select option {
-    color: black;
-}
-
-/* Submit Button Styles */
-
-/* save and submit button */
-input[type="button"] {
-    position: absolute;
+/* Save Button*/
+#save {
+    position: static;
     transition: background-color 1s ease; /* change of color takes a second(ex. when you hover) */
     background-color: #1565C0;
     /*navy blue*/
@@ -203,57 +158,38 @@ input[type="button"] {
     font-size: clamp(15px, 1.5vw, 30px);
 }
 
-/* save button */
-#save {
-    margin-left: 320px;
-    margin-top: -75px;
+/* Redo Button*/
+#redo {
+    position: static;
+    transition: background-color 1s ease; /* change of color takes a second(ex. when you hover) */
+    background-color: #1565C0;
+    /*navy blue*/
+    color: white;
+    padding: 10px 20px;
+    border: none;
+    border-radius: 4px;
+    cursor: pointer;
+    font-size: clamp(15px, 1.5vw, 30px);
 }
 
-/* save button when hovered by mouse */
-input[type="button"]:hover:enabled {
+/* save and redo button when hovered by mouse */
+#save:hover,
+#redo:hover {
     background-color: #0D47A1;
     /*midnight blue*/
 }
 
-/* save button when clicked */
-input[type="button"]:active:enabled {
+/* save and redo button when clicked */
+#save:active,
+#redo:active {
     background-color: #2196F3;
     /*light blue*/
 }
 
-/* save button when 'tab' to */
-input[type="button"]:focus:enabled {
+/* save and redo button when 'tab' to */
+#save:focus,
+#redo:active {
     outline: none;
     box-shadow: 0 0 0 2px #1A237E;
     /*dark blue*/
-}
-
-/* save button when disabled */
-input[type="button"]:disabled {
-    opacity: 0.5;
-}
-
-/* positioning for date */
-label[for="birthday"],
-input[type="date"] select {
-    margin-left: -340px;
-}
-
-/* positioning for category */
-input[type="category"],
-label[for="category"] select {
-    margin-left: -340px;
-}
-
-/* 'Birthday' and 'Select a category' labels */
-label {
-    font-family: 'pacifico';
-}
-
-/* font for dropdown menu on category */
-#category select{
-    font-size: 24px;
-    /* Adjust the font size as desired */
-    padding: 10px 16px;
-    /* Adjust the padding as desired */
 }
