--- source/horoscope/fortunePage.html
@@ -1,36 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <title>Horoscope App, Team #5</title>
-    <link rel="icon" type="image/x-icon" href="images/star_favicon.ico">
-    <link rel="stylesheet" type="text/css" href="./css/fortunePage.css">
-    <script src="./js/fortunePage.js" type="module"></script>
-</head>
-
-<body>
-    <main>
-        <video src="https://github.com/ZhouYuantian/CSE110-Storge/raw/main/background.mp4" id="bgvideo" autoplay
-            muted></video>
-        <img src="https://github.com/ZhouYuantian/CSE110-Storge/blob/main/background.jpg?raw=true" id="vbgimg">
-        <div id="main">
-            <form id="user-form">
-                <div id="save-redo-buttons" class="button-container">
-                    <div class="r-button">
-                        <input type="button" id="redo" value="REDO">
-                    </div>
-                    <div class="s-button">
-                        <input type="button" id="save" value="SAVE">
-                    </div>
-                </div>
-            </form>
-            <div id="output">
-                <h1 id="horoscope-title">Your Horoscope</h1>
-                <div id="horoscope-fortune">
-                </div>
-            </div>
-        </div>
-    </main>
-</body>
-
-</html>
